audio-model/requirements.txt
openai>=1.12.0
python-dotenv>=1.0.0
pandas>=2.0.0
anthropic>=0.18.0
metapub>=0.3.0
reportlab>=4.0.0
PyPDF2>=3.0.0
PyMuPDF>=1.23.0
flask>=3.0.0
